+++ docs/demos/cascader/PureJsx.zh-CN.tsx
@@ -1,4 +1,9 @@
-import { Cascader, FormButtonGroup, FormItem, Submit } from '@formily/antd-v5'
+import {
+  Cascader,
+  FormButtonGroup,
+  FormItem,
+  Submit,
+} from '@mozhou-tech/formily-adaptor-mo'
 import {
   createForm,
   Field as FieldType,