+++ docs/demos/checkbox/PureJsx.tsx
@@ -1,4 +1,9 @@
-import { Checkbox, FormButtonGroup, FormItem, Submit } from '@formily/antd-v5'
+import {
+  Checkbox,
+  FormButtonGroup,
+  FormItem,
+  Submit,
+} from '@mozhou-tech/formily-adaptor-mo'
 import {
   createForm,
   Field as FieldType,