+++ docs/demos/date-picker/Schema.zh-CN.tsx
@@ -1,4 +1,9 @@
-import { DatePicker, FormButtonGroup, FormItem, Submit } from '@formily/antd-v5'
+import {
+  DatePicker,
+  FormButtonGroup,
+  FormItem,
+  Submit,
+} from '@mozhou-tech/formily-adaptor-mo'
 import { createForm } from '@formily/core'
 import { createSchemaField, FormProvider, ISchema } from '@formily/react'
 import React from 'react'