+++ docs/demos/editable/Markup.zh-CN.tsx
@@ -5,7 +5,7 @@ import {
   FormItem,
   Input,
   Submit,
-} from '@formily/antd-v5'
+} from '@mozhou-tech/formily-adaptor-mo'
 import { createForm, ObjectField } from '@formily/core'
 import { createSchemaField, FormProvider } from '@formily/react'
 import React from 'react'