+++ docs/demos/form-grid/Native.zh-CN.tsx
@@ -1,4 +1,4 @@
-import { FormGrid } from '@formily/antd-v5'
+import { FormGrid } from '@mozhou-tech/formily-adaptor-mo'
 import React from 'react'
 
 const { GridColumn } = FormGrid