+++ docs/demos/form/Form.zh-CN.tsx
@@ -7,7 +7,7 @@ import {
   FormGrid,
   FormButtonGroup,
   Submit,
-} from '@formily/antd-v5'
+} from '@mozhou-tech/formily-adaptor-mo'
 import { createForm } from '@formily/core'
 import { Field } from '@formily/react'