+++ docs/demos/radio/PureJsx.zh-CN.tsx
@@ -1,4 +1,9 @@
-import { FormButtonGroup, FormItem, Radio, Submit } from '@formily/antd-v5'
+import {
+  FormButtonGroup,
+  FormItem,
+  Radio,
+  Submit,
+} from '@mozhou-tech/formily-adaptor-mo'
 import { createForm } from '@formily/core'
 import { Field, FormProvider } from '@formily/react'
 import React from 'react'