+++ docs/demos/reset/Verify.tsx
@@ -1,5 +1,10 @@
 import React from 'react'
-import { Input, FormItem, FormButtonGroup, Reset } from '@formily/antd-v5'
+import {
+  Input,
+  FormItem,
+  FormButtonGroup,
+  Reset,
+} from '@mozhou-tech/formily-adaptor-mo'
 import { createForm } from '@formily/core'
 import { FormProvider, createSchemaField } from '@formily/react'