+++ docs/demos/select/MarkupAsync.zh-CN.tsx
@@ -1,4 +1,9 @@
-import { FormButtonGroup, FormItem, Select, Submit } from '@formily/antd-v5'
+import {
+  FormButtonGroup,
+  FormItem,
+  Select,
+  Submit,
+} from '@mozhou-tech/formily-adaptor-mo'
 import {
   createForm,
   Field,