+++ docs/demos/time-picker/PureJsx.tsx
@@ -1,4 +1,9 @@
-import { FormButtonGroup, FormItem, Submit, TimePicker } from '@formily/antd-v5'
+import {
+  FormButtonGroup,
+  FormItem,
+  Submit,
+  TimePicker,
+} from '@mozhou-tech/formily-adaptor-mo'
 import { createForm } from '@formily/core'
 import { Field, FormProvider } from '@formily/react'
 import React from 'react'